backend/pyproject.toml
[tool.poetry]
name = "sentinel-backend"
version = "0.1.0"
description = "Sentinel — AI System Monitoring Platform"
authors = ["Sentinel Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
# Web framework
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
pydantic = {extras = ["email"], version = "^2.7"}
pydantic-settings = "^2.3"
python-multipart = "^0.0.9"

# Auth + security
python-jose = {extras = ["cryptography"], version = "^3.3"}
bcrypt = "3.2.2"
passlib = {extras = ["bcrypt"], version = "^1.7"}
slowapi = "^0.1.9"

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"

# Cache + messaging
redis = {extras = ["asyncio"], version = "^5.0"}
kafka-python = "^2.0"

# HTTP + LLM
httpx = "^0.27"
openai = "^1.35"

# ML — anomaly detection
scikit-learn = "^1.5"
numpy = "^1.26"
pandas = "^2.2"

# ML — deep learning LSTM autoencoder
tensorflow = "^2.16"
keras = "^3.3"

# ML — forecasting
prophet = "^1.1"
statsmodels = "^0.14"

# ML — graph analysis
networkx = "^3.3"

# ML — experiment tracking
mlflow = "^2.13"

# Observability
prometheus-fastapi-instrumentator = "^7.0"
opentelemetry-api = "^1.25"
opentelemetry-sdk = "^1.25"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-exporter-jaeger = "^1.21"
opentelemetry-instrumentation-sqlalchemy = "^0.46b0"
opentelemetry-instrumentation-redis = "^0.46b0"

# Utilities
apscheduler = "^3.10"
python-dotenv = "^1.0"
email-validator = "^2.1"
structlog = "^24.2"
setuptools = "^82.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
httpx = "^0.27"
ruff = "^0.4"
mypy = "^1.10"
pytest-cov = "^5.0"

[tool.ruff]
line-length = 88
select = ["E","F","I","N","W","UP"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
